Updates to barseq acquisition - turning into 3 files
